Draft: split clone from Draft.py
.


.
